Describe your data cleanly and effectively

Describe data sets with multiple variable types effectively.

describe_all(data, digits = 2, include_NAcat = TRUE, max_levels = 10,
  include_numeric = FALSE, sort_by_freq = FALSE, NAcat_include = NULL,
  ...)

describe_all_num(data, digits = 2, ...)

describe_all_cat(data, digits = 2, include_NAcat = TRUE, max_levels = 10,
  include_numeric = FALSE, sort_by_freq = FALSE)

Arguments

data
The dataset, of class data.frame.

digits
See round. Default is 2, which for categorical is applied to the proportion (i.e. before converting to percentage).

include_NAcat
Include NA values as categorical levels? Default is TRUE.

max_levels
The maximum number of levels you want to display for categorical variables. Default is 10.

include_numeric
For categorical summary, also include numeric variables with fewer or equal max_levels? Default is FALSE.

sort_by_freq
Sort categorical result by frequency? Default is FALSE.

NAcat_include
Deprecated include_NAcat.

...
Additional arguments passed to num_summary

Value

A list with two elements of summaries for numeric and other variables respectively.

Details

This function comes out of my frustrations from various data set summaries either being inadequate for my needs, too 'busy' with output, or unable to deal well with mixed data types. Numeric data is treated separately from categorical, and provides the same information as in num_summary. Categorical variables are defined as anything with equal or fewer distinct values than max_levels combined with include_numeric. Categorical variables are summarized with frequencies and percentages. For empty categorical variables (e.g. after a subset), a warning is thrown.

The functions describe_all_num and describe_all_cat will provide only numeric or only categorical data summaries respectively. describeAll is a deprecated alias.

See also

Examples

library(tidyext); library(dplyr)
X = data.frame(f1 =gl(2, 1, 20, labels=c('A', 'B')),
               f2=gl(2, 2, 20, labels=c('X', 'Q')))
X = X %>% mutate(bin1 = rbinom(20, 1, p=.5),
                 logic1 = sample(c(TRUE, FALSE), 20, replace = TRUE),
                 num1 = rnorm(20),
                 num2 = rpois(20, 5),
                 char1 = sample(letters, 20, replace = TRUE))
describe_all(X)

#> $`Numeric Variables`
#>   Variable  N Mean   SD   Min    Q1 Median   Q3   Max Missing
#> 1     bin1 20 0.30 0.47  0.00  0.00   0.00 1.00  1.00       0
#> 2     num1 20 0.01 0.99 -1.91 -0.44   0.05 0.54  1.89       0
#> 3     num2 20 4.95 2.42  1.00  3.00   5.00 6.00 12.00       0
#> 
#> $`Categorical Variables`
#>   Variable Group Frequency  %
#> 1       f1     A        10 50
#> 2       f1     B        10 50
#> 3       f2     Q        10 50
#> 4       f2     X        10 50
#> 5   logic1 FALSE        10 50
#> 6   logic1  TRUE        10 50
#> 

describe_all(data.frame(x=factor(1:7)), digits=5)

#> No numeric data.

#> $`Numeric Variables`
#> NULL
#> 
#> $`Categorical Variables`
#>   Variable Group Frequency      %
#> 1        x     1         1 14.286
#> 2        x     2         1 14.286
#> 3        x     3         1 14.286
#> 4        x     4         1 14.286
#> 5        x     5         1 14.286
#> 6        x     6         1 14.286
#> 7        x     7         1 14.286
#>