Summarize data by groups

Provide common numeric summary information.

num_by(data, main_var, group_var, digits = 1, extra = FALSE)

cat_by(data, main_var, group_var, digits = FALSE, perc_by_group = TRUE,
  sort_by_group = TRUE)

Arguments

data
data frame

main_var
the variable to be summarized. Multiple variables are supplied using the vars() function.

group_var
the (optional) grouping variable.

digits
Optional rounding. Default is 1.

extra
See num_summary.

perc_by_group
when supplied a grouping variable for cat_by, do you want within group percentages also (default is TRUE)

sort_by_group
when supplied a grouping variable for cat_by, do you want result sorted on the grouping variable? Default is TRUE.

Value

data.frame/tibble with the corresponding summary statistics

Details

The num_by function takes a numeric variable from a dataframe and provides sample size, mean, standard deviation, min, first quartile, median, third quartile, max, and number of missing values, possibly over a grouping variable.

It works in the dplyr style using unquoted (bare) variable names, using the vars() function if there is more than one variable. If using a grouping variable, it will treat missing values as a separate group.

For cat_by, frequencies and percentage (out of total or group_var) are returned. Warnings are given if any of the main variables are more than 10 levels, or the data appears to be non-categorical. The group_var argument is essentially just to provide percentages based on a focal grouping variable out of all those supplied. Missing values are treated as an additional unique value.

Missing values are treated as separate groups, as it's often useful to explore the nature of the missingness. To avoid this, just use na.omit or dplyr::drop_na on the data frame first.

See also

Examples

library(tidyext)
df1 <- data.frame(g1 = factor(sample(1:2, 50, replace = TRUE), labels=c('a','b')),
                  g2 = sample(1:4, 50, replace = TRUE),
                  a = rnorm(50),
                  b = rpois(50, 10),
                  c = sample(letters, 50, replace=TRUE),
                  d = sample(c(TRUE, FALSE), 50, replace=TRUE)
                 )

num_by(df1, main_var = a)

#>   Variable  N Mean  SD  Min   Q1 Median  Q3 Max Missing
#> 1        a 50 -0.1 1.1 -3.7 -0.6   -0.1 0.7 1.9       0

num_by(df1, main_var = a, group_var = g2, digits=2)

#> # A tibble: 4 x 11
#>      g2 Variable     N  Mean    SD   Min    Q1 Median     Q3   Max Missing
#>   <int> <chr>    <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl>   <dbl>
#> 1     1 a           13 -0.74  1.27 -3.71 -0.78  -0.61 -0.290  1.06       0
#> 2     2 a           13  0.37  1.26 -2.29 -0.31   0.68  1.57   1.86       0
#> 3     3 a           11 -0.14  0.87 -1.77 -0.71  -0.07  0.6    1.18       0
#> 4     4 a           13  0.1   0.67 -1.29 -0.25   0.01  0.38   1.2        0

num_by(df1, main_var = dplyr::vars(a,b), group_var = g1, digits=1)

#> # A tibble: 4 x 11
#>   g1    Variable     N  Mean    SD   Min    Q1 Median    Q3   Max Missing
#>   <fct> <chr>    <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl>   <dbl>
#> 1 a     a           20  -0.3   0.9  -2.3  -0.8   -0.3   0.4   1.2       0
#> 2 a     b           20   9.8   3.1   6     7     10    11.5  16         0
#> 3 b     a           30   0     1.2  -3.7  -0.5    0     0.8   1.9       0
#> 4 b     b           30  10.2   3.9   2     8     10    13    17         0

cat_by(df1, main_var = g1, group_var = g2, digits=1)

#> `mutate_if()` ignored the following grouping variables:
#> Column `g2`

#> # A tibble: 8 x 5
#> # Groups:   g2 [4]
#>      g2 g1        N `% of Total` `% of g2`
#>   <int> <fct> <dbl>        <dbl>     <dbl>
#> 1     1 a         4            8      30.8
#> 2     1 b         9           18      69.2
#> 3     2 a         5           10      38.5
#> 4     2 b         8           16      61.5
#> 5     3 a         7           14      63.6
#> 6     3 b         4            8      36.4
#> 7     4 a         4            8      30.8
#> 8     4 b         9           18      69.2

cat_by(df1, main_var = dplyr::vars(g1,d), group_var = g2, perc_by_group=FALSE)

#> # A tibble: 15 x 5
#>       g2 g1    d         N `% of Total`
#>    <int> <fct> <lgl> <int>        <dbl>
#>  1     1 a     FALSE     1            2
#>  2     1 a     TRUE      3            6
#>  3     1 b     FALSE     2            4
#>  4     1 b     TRUE      7           14
#>  5     2 a     FALSE     3            6
#>  6     2 a     TRUE      2            4
#>  7     2 b     FALSE     2            4
#>  8     2 b     TRUE      6           12
#>  9     3 a     FALSE     3            6
#> 10     3 a     TRUE      4            8
#> 11     3 b     FALSE     3            6
#> 12     3 b     TRUE      1            2
#> 13     4 a     TRUE      4            8
#> 14     4 b     FALSE     5           10
#> 15     4 b     TRUE      4            8