Perform common preprocessing of numeric data

Perform common data transformations

pre_process(data, std = TRUE, scale_by = 1, log_vars = NULL,
  log_base = exp(1), zero_start = NULL, zero_one = NULL)

Arguments

data
A data frame or tibble.

std
Standardize numeric/integer variables?

scale_by
A single value to standardize by. See details. Default is 1.

log_vars
Which variables to log. Requires vars().

log_base
Log base. Default is exp(1).

zero_start
Which variables to start by zero. Requires vars().

zero_one
Which variables to rescale from 0 to 1. Requires vars().

Value

A data frame that has been processed

Details

At a minimum, by default, this function will standardize numeric/integer variables in a data set by the value provided to std (1 standard deviation is the default for scale_by). If scale_by is set to zero, the variables will simply be centered and not scaled. This operation will only be performed on variables not provided for the other options.

- It will log variables contained within vars(), with base equal to log_base, which is exp(1) by default (i.e. natural log).

- zero_start will make the minimum value zero, as often done for time index variables in longitudinal data. - zero_one will rescale variables to range from 0 to 1.

This is a bare minimum function, meant to perform common operations quickly/easily. If you're wanting to do more than this, you'll have to do it yourself.

I would have called this function transmute, but it's already taken by dplyr, despite no one ever using it for the purpose of turning into a whirlwind capable of taking out whatever the Galactor may have in store for them.

Examples

library(tidyext)
pre_process(mtcars)

#>            mpg        cyl        disp          hp        drat           wt
#> 1   0.15088482 -0.1049878 -0.57061982 -0.53509284  0.56751369 -0.610399567
#> 2   0.15088482 -0.1049878 -0.57061982 -0.53509284  0.56751369 -0.349785269
#> 3   0.44954345 -1.2248578 -0.99018209 -0.78304046  0.47399959 -0.917004624
#> 4   0.21725341 -0.1049878  0.22009369 -0.53509284 -0.96611753 -0.002299538
#> 5  -0.23073453  1.0148821  1.04308123  0.41294217 -0.83519779  0.227654255
#> 6  -0.33028740 -0.1049878 -0.04616698 -0.60801861 -1.56460776  0.248094592
#> 7  -0.96078893  1.0148821  1.04308123  1.43390296 -0.72298087  0.360516446
#> 8   0.71501778 -1.2248578 -0.67793094 -1.23518023  0.17475447 -0.027849959
#> 9   0.44954345 -1.2248578 -0.72553512 -0.75387015  0.60491932 -0.068730634
#> 10 -0.14777380 -0.1049878 -0.50929918 -0.34548584  0.60491932  0.227654255
#> 11 -0.38006384 -0.1049878 -0.50929918 -0.34548584  0.60491932  0.227654255
#> 12 -0.61235388  1.0148821  0.36371309  0.48586794 -0.98482035  0.871524874
#> 13 -0.46302456  1.0148821  0.36371309  0.48586794 -0.98482035  0.524039143
#> 14 -0.81145962  1.0148821  0.36371309  0.48586794 -0.98482035  0.575139986
#> 15 -1.60788262  1.0148821  1.94675381  0.85049680 -1.24665983  2.077504765
#> 16 -1.60788262  1.0148821  1.84993175  0.99634834 -1.11574009  2.255335698
#> 17 -0.89442035  1.0148821  1.68856165  1.21512565 -0.68557523  2.174596366
#> 18  2.04238943 -1.2248578 -1.22658929 -1.17683962  0.90416444 -1.039646647
#> 19  1.71054652 -1.2248578 -1.25079481 -1.38103178  2.49390411 -1.637526508
#> 20  2.29127162 -1.2248578 -1.28790993 -1.19142477  1.16600392 -1.412682800
#> 21  0.23384555 -1.2248578 -0.89255318 -0.72469984  0.19345729 -0.768812180
#> 22 -0.76168319  1.0148821  0.70420401  0.04831332 -1.56460776  0.309415603
#> 23 -0.81145962  1.0148821  0.59124494  0.04831332 -0.83519779  0.222544170
#> 24 -1.12671039  1.0148821  0.96239618  1.43390296  0.24956575  0.636460997
#> 25 -0.14777380  1.0148821  1.36582144  0.41294217 -0.96611753  0.641571082
#> 26  1.19619000 -1.2248578 -1.22416874 -1.17683962  0.90416444 -1.310481114
#> 27  0.98049211 -1.2248578 -0.89093948 -0.81221077  1.55876313 -1.100967659
#> 28  1.71054652 -1.2248578 -1.09426581 -0.49133738  0.32437703 -1.741772228
#> 29 -0.71190675  1.0148821  0.97046468  1.71102089  1.16600392 -0.048290296
#> 30 -0.06481307 -0.1049878 -0.69164740  0.41294217  0.04383473 -0.457097039
#> 31 -0.84464392  1.0148821  0.56703942  2.74656682 -0.10578782  0.360516446
#> 32  0.21725341 -1.2248578 -0.88529152 -0.54967799  0.96027290 -0.446876870
#>           qsec         vs         am       gear       carb
#> 1  -0.77716515 -0.8680278  1.1899014  0.4235542  0.7352031
#> 2  -0.46378082 -0.8680278  1.1899014  0.4235542  0.7352031
#> 3   0.42600682  1.1160357  1.1899014  0.4235542 -1.1221521
#> 4   0.89048716  1.1160357 -0.8141431 -0.9318192 -1.1221521
#> 5  -0.46378082 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 6   1.32698675  1.1160357 -0.8141431 -0.9318192 -1.1221521
#> 7  -1.12412636 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 8   1.20387148  1.1160357 -0.8141431  0.4235542 -0.5030337
#> 9   2.82675459  1.1160357 -0.8141431  0.4235542 -0.5030337
#> 10  0.25252621  1.1160357 -0.8141431  0.4235542  0.7352031
#> 11  0.58829513  1.1160357 -0.8141431  0.4235542  0.7352031
#> 12 -0.25112717 -0.8680278 -0.8141431 -0.9318192  0.1160847
#> 13 -0.13920420 -0.8680278 -0.8141431 -0.9318192  0.1160847
#> 14  0.08464175 -0.8680278 -0.8141431 -0.9318192  0.1160847
#> 15  0.07344945 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 16 -0.01608893 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 17 -0.23993487 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 18  0.90727560  1.1160357  1.1899014  0.4235542 -1.1221521
#> 19  0.37564148  1.1160357  1.1899014  0.4235542 -0.5030337
#> 20  1.14790999  1.1160357  1.1899014  0.4235542 -1.1221521
#> 21  1.20946763  1.1160357 -0.8141431 -0.9318192 -1.1221521
#> 22 -0.54772305 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 23 -0.30708866 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 24 -1.36476075 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 25 -0.44699237 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 26  0.58829513  1.1160357  1.1899014  0.4235542 -1.1221521
#> 27 -0.64285758 -0.8680278  1.1899014  1.7789276 -0.5030337
#> 28 -0.53093460  1.1160357  1.1899014  1.7789276 -0.5030337
#> 29 -1.87401028 -0.8680278  1.1899014  1.7789276  0.7352031
#> 30 -1.31439542 -0.8680278  1.1899014  1.7789276  1.9734398
#> 31 -1.81804880 -0.8680278  1.1899014  1.7789276  3.2116766
#> 32  0.42041067  1.1160357  1.1899014  0.4235542 -0.5030337

pre_process(mtcars, log_vars=vars(mpg, wt))

#>         mpg        cyl        disp          hp        drat        wt
#> 1  3.044522 -0.1049878 -0.57061982 -0.53509284  0.56751369 0.9631743
#> 2  3.044522 -0.1049878 -0.57061982 -0.53509284  0.56751369 1.0560527
#> 3  3.126761 -1.2248578 -0.99018209 -0.78304046  0.47399959 0.8415672
#> 4  3.063391 -0.1049878  0.22009369 -0.53509284 -0.96611753 1.1678274
#> 5  2.928524  1.0148821  1.04308123  0.41294217 -0.83519779 1.2354715
#> 6  2.895912 -0.1049878 -0.04616698 -0.60801861 -1.56460776 1.2412686
#> 7  2.660260  1.0148821  1.04308123  1.43390296 -0.72298087 1.2725656
#> 8  3.194583 -1.2248578 -0.67793094 -1.23518023  0.17475447 1.1600209
#> 9  3.126761 -1.2248578 -0.72553512 -0.75387015  0.60491932 1.1474025
#> 10 2.954910 -0.1049878 -0.50929918 -0.34548584  0.60491932 1.2354715
#> 11 2.879198 -0.1049878 -0.50929918 -0.34548584  0.60491932 1.2354715
#> 12 2.797281  1.0148821  0.36371309  0.48586794 -0.98482035 1.4036430
#> 13 2.850707  1.0148821  0.36371309  0.48586794 -0.98482035 1.3164082
#> 14 2.721295  1.0148821  0.36371309  0.48586794 -0.98482035 1.3297240
#> 15 2.341806  1.0148821  1.94675381  0.85049680 -1.24665983 1.6582281
#> 16 2.341806  1.0148821  1.84993175  0.99634834 -1.11574009 1.6908336
#> 17 2.687847  1.0148821  1.68856165  1.21512565 -0.68557523 1.6761615
#> 18 3.478158 -1.2248578 -1.22658929 -1.17683962  0.90416444 0.7884574
#> 19 3.414443 -1.2248578 -1.25079481 -1.38103178  2.49390411 0.4793350
#> 20 3.523415 -1.2248578 -1.28790993 -1.19142477  1.16600392 0.6070445
#> 21 3.068053 -1.2248578 -0.89255318 -0.72469984  0.19345729 0.9021918
#> 22 2.740840  1.0148821  0.70420401  0.04831332 -1.56460776 1.2584610
#> 23 2.721295  1.0148821  0.59124494  0.04831332 -0.83519779 1.2340169
#> 24 2.587764  1.0148821  0.96239618  1.43390296  0.24956575 1.3454724
#> 25 2.954910  1.0148821  1.36582144  0.41294217 -0.96611753 1.3467736
#> 26 3.306887 -1.2248578 -1.22416874 -1.17683962  0.90416444 0.6601073
#> 27 3.258097 -1.2248578 -0.89093948 -0.81221077  1.55876313 0.7608058
#> 28 3.414443 -1.2248578 -1.09426581 -0.49133738  0.32437703 0.4140944
#> 29 2.760010  1.0148821  0.97046468  1.71102089  1.16600392 1.1537316
#> 30 2.980619 -0.1049878 -0.69164740  0.41294217  0.04383473 1.0188473
#> 31 2.708050  1.0148821  0.56703942  2.74656682 -0.10578782 1.2725656
#> 32 3.063391 -1.2248578 -0.88529152 -0.54967799  0.96027290 1.0224509
#>           qsec         vs         am       gear       carb
#> 1  -0.77716515 -0.8680278  1.1899014  0.4235542  0.7352031
#> 2  -0.46378082 -0.8680278  1.1899014  0.4235542  0.7352031
#> 3   0.42600682  1.1160357  1.1899014  0.4235542 -1.1221521
#> 4   0.89048716  1.1160357 -0.8141431 -0.9318192 -1.1221521
#> 5  -0.46378082 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 6   1.32698675  1.1160357 -0.8141431 -0.9318192 -1.1221521
#> 7  -1.12412636 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 8   1.20387148  1.1160357 -0.8141431  0.4235542 -0.5030337
#> 9   2.82675459  1.1160357 -0.8141431  0.4235542 -0.5030337
#> 10  0.25252621  1.1160357 -0.8141431  0.4235542  0.7352031
#> 11  0.58829513  1.1160357 -0.8141431  0.4235542  0.7352031
#> 12 -0.25112717 -0.8680278 -0.8141431 -0.9318192  0.1160847
#> 13 -0.13920420 -0.8680278 -0.8141431 -0.9318192  0.1160847
#> 14  0.08464175 -0.8680278 -0.8141431 -0.9318192  0.1160847
#> 15  0.07344945 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 16 -0.01608893 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 17 -0.23993487 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 18  0.90727560  1.1160357  1.1899014  0.4235542 -1.1221521
#> 19  0.37564148  1.1160357  1.1899014  0.4235542 -0.5030337
#> 20  1.14790999  1.1160357  1.1899014  0.4235542 -1.1221521
#> 21  1.20946763  1.1160357 -0.8141431 -0.9318192 -1.1221521
#> 22 -0.54772305 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 23 -0.30708866 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 24 -1.36476075 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 25 -0.44699237 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 26  0.58829513  1.1160357  1.1899014  0.4235542 -1.1221521
#> 27 -0.64285758 -0.8680278  1.1899014  1.7789276 -0.5030337
#> 28 -0.53093460  1.1160357  1.1899014  1.7789276 -0.5030337
#> 29 -1.87401028 -0.8680278  1.1899014  1.7789276  0.7352031
#> 30 -1.31439542 -0.8680278  1.1899014  1.7789276  1.9734398
#> 31 -1.81804880 -0.8680278  1.1899014  1.7789276  3.2116766
#> 32  0.42041067  1.1160357  1.1899014  0.4235542 -0.5030337

pre_process(mtcars, zero_start=vars(cyl, gear))

#>            mpg cyl        disp          hp        drat           wt        qsec
#> 1   0.15088482   2 -0.57061982 -0.53509284  0.56751369 -0.610399567 -0.77716515
#> 2   0.15088482   2 -0.57061982 -0.53509284  0.56751369 -0.349785269 -0.46378082
#> 3   0.44954345   0 -0.99018209 -0.78304046  0.47399959 -0.917004624  0.42600682
#> 4   0.21725341   2  0.22009369 -0.53509284 -0.96611753 -0.002299538  0.89048716
#> 5  -0.23073453   4  1.04308123  0.41294217 -0.83519779  0.227654255 -0.46378082
#> 6  -0.33028740   2 -0.04616698 -0.60801861 -1.56460776  0.248094592  1.32698675
#> 7  -0.96078893   4  1.04308123  1.43390296 -0.72298087  0.360516446 -1.12412636
#> 8   0.71501778   0 -0.67793094 -1.23518023  0.17475447 -0.027849959  1.20387148
#> 9   0.44954345   0 -0.72553512 -0.75387015  0.60491932 -0.068730634  2.82675459
#> 10 -0.14777380   2 -0.50929918 -0.34548584  0.60491932  0.227654255  0.25252621
#> 11 -0.38006384   2 -0.50929918 -0.34548584  0.60491932  0.227654255  0.58829513
#> 12 -0.61235388   4  0.36371309  0.48586794 -0.98482035  0.871524874 -0.25112717
#> 13 -0.46302456   4  0.36371309  0.48586794 -0.98482035  0.524039143 -0.13920420
#> 14 -0.81145962   4  0.36371309  0.48586794 -0.98482035  0.575139986  0.08464175
#> 15 -1.60788262   4  1.94675381  0.85049680 -1.24665983  2.077504765  0.07344945
#> 16 -1.60788262   4  1.84993175  0.99634834 -1.11574009  2.255335698 -0.01608893
#> 17 -0.89442035   4  1.68856165  1.21512565 -0.68557523  2.174596366 -0.23993487
#> 18  2.04238943   0 -1.22658929 -1.17683962  0.90416444 -1.039646647  0.90727560
#> 19  1.71054652   0 -1.25079481 -1.38103178  2.49390411 -1.637526508  0.37564148
#> 20  2.29127162   0 -1.28790993 -1.19142477  1.16600392 -1.412682800  1.14790999
#> 21  0.23384555   0 -0.89255318 -0.72469984  0.19345729 -0.768812180  1.20946763
#> 22 -0.76168319   4  0.70420401  0.04831332 -1.56460776  0.309415603 -0.54772305
#> 23 -0.81145962   4  0.59124494  0.04831332 -0.83519779  0.222544170 -0.30708866
#> 24 -1.12671039   4  0.96239618  1.43390296  0.24956575  0.636460997 -1.36476075
#> 25 -0.14777380   4  1.36582144  0.41294217 -0.96611753  0.641571082 -0.44699237
#> 26  1.19619000   0 -1.22416874 -1.17683962  0.90416444 -1.310481114  0.58829513
#> 27  0.98049211   0 -0.89093948 -0.81221077  1.55876313 -1.100967659 -0.64285758
#> 28  1.71054652   0 -1.09426581 -0.49133738  0.32437703 -1.741772228 -0.53093460
#> 29 -0.71190675   4  0.97046468  1.71102089  1.16600392 -0.048290296 -1.87401028
#> 30 -0.06481307   2 -0.69164740  0.41294217  0.04383473 -0.457097039 -1.31439542
#> 31 -0.84464392   4  0.56703942  2.74656682 -0.10578782  0.360516446 -1.81804880
#> 32  0.21725341   0 -0.88529152 -0.54967799  0.96027290 -0.446876870  0.42041067
#>            vs         am gear       carb
#> 1  -0.8680278  1.1899014    1  0.7352031
#> 2  -0.8680278  1.1899014    1  0.7352031
#> 3   1.1160357  1.1899014    1 -1.1221521
#> 4   1.1160357 -0.8141431    0 -1.1221521
#> 5  -0.8680278 -0.8141431    0 -0.5030337
#> 6   1.1160357 -0.8141431    0 -1.1221521
#> 7  -0.8680278 -0.8141431    0  0.7352031
#> 8   1.1160357 -0.8141431    1 -0.5030337
#> 9   1.1160357 -0.8141431    1 -0.5030337
#> 10  1.1160357 -0.8141431    1  0.7352031
#> 11  1.1160357 -0.8141431    1  0.7352031
#> 12 -0.8680278 -0.8141431    0  0.1160847
#> 13 -0.8680278 -0.8141431    0  0.1160847
#> 14 -0.8680278 -0.8141431    0  0.1160847
#> 15 -0.8680278 -0.8141431    0  0.7352031
#> 16 -0.8680278 -0.8141431    0  0.7352031
#> 17 -0.8680278 -0.8141431    0  0.7352031
#> 18  1.1160357  1.1899014    1 -1.1221521
#> 19  1.1160357  1.1899014    1 -0.5030337
#> 20  1.1160357  1.1899014    1 -1.1221521
#> 21  1.1160357 -0.8141431    0 -1.1221521
#> 22 -0.8680278 -0.8141431    0 -0.5030337
#> 23 -0.8680278 -0.8141431    0 -0.5030337
#> 24 -0.8680278 -0.8141431    0  0.7352031
#> 25 -0.8680278 -0.8141431    0 -0.5030337
#> 26  1.1160357  1.1899014    1 -1.1221521
#> 27 -0.8680278  1.1899014    2 -0.5030337
#> 28  1.1160357  1.1899014    2 -0.5030337
#> 29 -0.8680278  1.1899014    2  0.7352031
#> 30 -0.8680278  1.1899014    2  1.9734398
#> 31 -0.8680278  1.1899014    2  3.2116766
#> 32  1.1160357  1.1899014    1 -0.5030337

pre_process(mtcars, zero_one=vars(mpg))

#>          mpg        cyl        disp          hp        drat           wt
#> 1  0.4510638 -0.1049878 -0.57061982 -0.53509284  0.56751369 -0.610399567
#> 2  0.4510638 -0.1049878 -0.57061982 -0.53509284  0.56751369 -0.349785269
#> 3  0.5276596 -1.2248578 -0.99018209 -0.78304046  0.47399959 -0.917004624
#> 4  0.4680851 -0.1049878  0.22009369 -0.53509284 -0.96611753 -0.002299538
#> 5  0.3531915  1.0148821  1.04308123  0.41294217 -0.83519779  0.227654255
#> 6  0.3276596 -0.1049878 -0.04616698 -0.60801861 -1.56460776  0.248094592
#> 7  0.1659574  1.0148821  1.04308123  1.43390296 -0.72298087  0.360516446
#> 8  0.5957447 -1.2248578 -0.67793094 -1.23518023  0.17475447 -0.027849959
#> 9  0.5276596 -1.2248578 -0.72553512 -0.75387015  0.60491932 -0.068730634
#> 10 0.3744681 -0.1049878 -0.50929918 -0.34548584  0.60491932  0.227654255
#> 11 0.3148936 -0.1049878 -0.50929918 -0.34548584  0.60491932  0.227654255
#> 12 0.2553191  1.0148821  0.36371309  0.48586794 -0.98482035  0.871524874
#> 13 0.2936170  1.0148821  0.36371309  0.48586794 -0.98482035  0.524039143
#> 14 0.2042553  1.0148821  0.36371309  0.48586794 -0.98482035  0.575139986
#> 15 0.0000000  1.0148821  1.94675381  0.85049680 -1.24665983  2.077504765
#> 16 0.0000000  1.0148821  1.84993175  0.99634834 -1.11574009  2.255335698
#> 17 0.1829787  1.0148821  1.68856165  1.21512565 -0.68557523  2.174596366
#> 18 0.9361702 -1.2248578 -1.22658929 -1.17683962  0.90416444 -1.039646647
#> 19 0.8510638 -1.2248578 -1.25079481 -1.38103178  2.49390411 -1.637526508
#> 20 1.0000000 -1.2248578 -1.28790993 -1.19142477  1.16600392 -1.412682800
#> 21 0.4723404 -1.2248578 -0.89255318 -0.72469984  0.19345729 -0.768812180
#> 22 0.2170213  1.0148821  0.70420401  0.04831332 -1.56460776  0.309415603
#> 23 0.2042553  1.0148821  0.59124494  0.04831332 -0.83519779  0.222544170
#> 24 0.1234043  1.0148821  0.96239618  1.43390296  0.24956575  0.636460997
#> 25 0.3744681  1.0148821  1.36582144  0.41294217 -0.96611753  0.641571082
#> 26 0.7191489 -1.2248578 -1.22416874 -1.17683962  0.90416444 -1.310481114
#> 27 0.6638298 -1.2248578 -0.89093948 -0.81221077  1.55876313 -1.100967659
#> 28 0.8510638 -1.2248578 -1.09426581 -0.49133738  0.32437703 -1.741772228
#> 29 0.2297872  1.0148821  0.97046468  1.71102089  1.16600392 -0.048290296
#> 30 0.3957447 -0.1049878 -0.69164740  0.41294217  0.04383473 -0.457097039
#> 31 0.1957447  1.0148821  0.56703942  2.74656682 -0.10578782  0.360516446
#> 32 0.4680851 -1.2248578 -0.88529152 -0.54967799  0.96027290 -0.446876870
#>           qsec         vs         am       gear       carb
#> 1  -0.77716515 -0.8680278  1.1899014  0.4235542  0.7352031
#> 2  -0.46378082 -0.8680278  1.1899014  0.4235542  0.7352031
#> 3   0.42600682  1.1160357  1.1899014  0.4235542 -1.1221521
#> 4   0.89048716  1.1160357 -0.8141431 -0.9318192 -1.1221521
#> 5  -0.46378082 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 6   1.32698675  1.1160357 -0.8141431 -0.9318192 -1.1221521
#> 7  -1.12412636 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 8   1.20387148  1.1160357 -0.8141431  0.4235542 -0.5030337
#> 9   2.82675459  1.1160357 -0.8141431  0.4235542 -0.5030337
#> 10  0.25252621  1.1160357 -0.8141431  0.4235542  0.7352031
#> 11  0.58829513  1.1160357 -0.8141431  0.4235542  0.7352031
#> 12 -0.25112717 -0.8680278 -0.8141431 -0.9318192  0.1160847
#> 13 -0.13920420 -0.8680278 -0.8141431 -0.9318192  0.1160847
#> 14  0.08464175 -0.8680278 -0.8141431 -0.9318192  0.1160847
#> 15  0.07344945 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 16 -0.01608893 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 17 -0.23993487 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 18  0.90727560  1.1160357  1.1899014  0.4235542 -1.1221521
#> 19  0.37564148  1.1160357  1.1899014  0.4235542 -0.5030337
#> 20  1.14790999  1.1160357  1.1899014  0.4235542 -1.1221521
#> 21  1.20946763  1.1160357 -0.8141431 -0.9318192 -1.1221521
#> 22 -0.54772305 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 23 -0.30708866 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 24 -1.36476075 -0.8680278 -0.8141431 -0.9318192  0.7352031
#> 25 -0.44699237 -0.8680278 -0.8141431 -0.9318192 -0.5030337
#> 26  0.58829513  1.1160357  1.1899014  0.4235542 -1.1221521
#> 27 -0.64285758 -0.8680278  1.1899014  1.7789276 -0.5030337
#> 28 -0.53093460  1.1160357  1.1899014  1.7789276 -0.5030337
#> 29 -1.87401028 -0.8680278  1.1899014  1.7789276  0.7352031
#> 30 -1.31439542 -0.8680278  1.1899014  1.7789276  1.9734398
#> 31 -1.81804880 -0.8680278  1.1899014  1.7789276  3.2116766
#> 32  0.42041067  1.1160357  1.1899014  0.4235542 -0.5030337