Spread without frustration

Spread without frustration

spread2(
  data,
  key,
  value,
  compact = TRUE,
  fill = NA,
  convert = FALSE,
  drop = TRUE,
  sep = NULL
)

Arguments

data
See spread.

key
See spread.

value
See spread.

compact
When you don't have a unique identifier there is more than one way the output could be produced. Actually two. One is rather messy but may in fact be what is desired. The other is more 'compact'. See the examples.

fill
See spread.

convert
See spread.

drop
See spread.

sep
See spread.

Value

A data frame with 'wide' format.

Details

Anyone that uses spread more than once has likely come across the duplicate identifiers problem. This function works just like select (it's mostly just a wrapper) except that it works when you don't have unique identifiers, which is actually quite often. The only additional functionality regards the compact argument. See the examples for what this means.

For more, see StackOverflow for example, or the numerous repeated issue brought up at the tidyr GitHub issues page. Search for 'spread duplicate' for all, not just open issues (direct link couldn't be handled by R documentation).

See also

Examples

library(tidyext); library(tidyr)

# initial example from spread
stocks_init <- data.frame(
time = as.Date('2009-01-01') + 0:9,
X = rnorm(10, 0, 1),
Y = rnorm(10, 0, 2),
Z = rnorm(10, 0, 4)
)

# a very common situation
stocks <- data.frame(
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4)
)

stocksm_init <- stocks_init %>% gather(stock, price, -time)
stocksm_init %>% spread(stock, price)    # works fine

#>          time          X          Y          Z
#> 1  2009-01-01  1.2240818 -2.1356474  1.7058569
#> 2  2009-01-02  0.3598138 -0.4359498 -1.1802859
#> 3  2009-01-03  0.4007715 -2.0520089  3.5805026
#> 4  2009-01-04  0.1106827 -1.4577825  3.5125340
#> 5  2009-01-05 -0.5558411 -1.2500785  3.2863243
#> 6  2009-01-06  1.7869131 -3.3733866  2.7545610
#> 7  2009-01-07  0.4978505  1.6755741  2.2156706
#> 8  2009-01-08 -1.9666172  0.3067462 -0.2476468
#> 9  2009-01-09  0.7013559 -2.2762739 -1.2238507
#> 10 2009-01-10 -0.4727914  2.5076298 -1.5218840

stocksm <- stocks %>% gather(stock, price)
# stocksm %>% spread(stock, price)       # annoying
stocksm %>% spread2(stock, price)        # works fine

#>              X           Y          Z
#> 1  -0.69470698  0.50663703  1.5185579
#> 2  -0.20791728 -0.05709351 -2.0092938
#> 3  -1.26539635 -0.08574091 -1.3328295
#> 4   2.16895597  2.73720457 -4.0743015
#> 5   1.20796200 -0.45154197 -4.2871649
#> 6  -1.12310858  3.03294121  1.2141146
#> 7  -0.40288484 -3.09750561  1.7928391
#> 8  -0.46665535  1.16922750  0.2120169
#> 9   0.77996512  0.24770849  3.6890699
#> 10 -0.08336907  0.43188314  8.2003387

stocksm %>% spread2(stock, price, compact = FALSE)

#>              X           Y          Z
#> 1  -0.69470698          NA         NA
#> 2  -0.20791728          NA         NA
#> 3  -1.26539635          NA         NA
#> 4   2.16895597          NA         NA
#> 5   1.20796200          NA         NA
#> 6  -1.12310858          NA         NA
#> 7  -0.40288484          NA         NA
#> 8  -0.46665535          NA         NA
#> 9   0.77996512          NA         NA
#> 10 -0.08336907          NA         NA
#> 11          NA  0.50663703         NA
#> 12          NA -0.05709351         NA
#> 13          NA -0.08574091         NA
#> 14          NA  2.73720457         NA
#> 15          NA -0.45154197         NA
#> 16          NA  3.03294121         NA
#> 17          NA -3.09750561         NA
#> 18          NA  1.16922750         NA
#> 19          NA  0.24770849         NA
#> 20          NA  0.43188314         NA
#> 21          NA          NA  1.5185579
#> 22          NA          NA -2.0092938
#> 23          NA          NA -1.3328295
#> 24          NA          NA -4.0743015
#> 25          NA          NA -4.2871649
#> 26          NA          NA  1.2141146
#> 27          NA          NA  1.7928391
#> 28          NA          NA  0.2120169
#> 29          NA          NA  3.6890699
#> 30          NA          NA  8.2003387

# works with NA
stocksm$price[sample(1:nrow(stocksm), 5)] = NA
stocksm %>% spread2(stock, price)

#>              X           Y         Z
#> 1  -0.69470698  0.50663703  1.518558
#> 2  -0.20791728          NA -2.009294
#> 3  -1.26539635 -0.08574091 -1.332830
#> 4   2.16895597          NA -4.074302
#> 5   1.20796200 -0.45154197 -4.287165
#> 6  -1.12310858  3.03294121  1.214115
#> 7  -0.40288484          NA  1.792839
#> 8  -0.46665535  1.16922750        NA
#> 9   0.77996512  0.24770849  3.689070
#> 10 -0.08336907          NA  8.200339

stocksm %>% spread2(stock, price, compact = FALSE)

#>              X           Y         Z
#> 1  -0.69470698          NA        NA
#> 2  -0.20791728          NA        NA
#> 3  -1.26539635          NA        NA
#> 4   2.16895597          NA        NA
#> 5   1.20796200          NA        NA
#> 6  -1.12310858          NA        NA
#> 7  -0.40288484          NA        NA
#> 8  -0.46665535          NA        NA
#> 9   0.77996512          NA        NA
#> 10 -0.08336907          NA        NA
#> 11          NA  0.50663703        NA
#> 12          NA          NA        NA
#> 13          NA -0.08574091        NA
#> 14          NA          NA        NA
#> 15          NA -0.45154197        NA
#> 16          NA  3.03294121        NA
#> 17          NA          NA        NA
#> 18          NA  1.16922750        NA
#> 19          NA  0.24770849        NA
#> 20          NA          NA        NA
#> 21          NA          NA  1.518558
#> 22          NA          NA -2.009294
#> 23          NA          NA -1.332830
#> 24          NA          NA -4.074302
#> 25          NA          NA -4.287165
#> 26          NA          NA  1.214115
#> 27          NA          NA  1.792839
#> 28          NA          NA        NA
#> 29          NA          NA  3.689070
#> 30          NA          NA  8.200339