Extend tidyr gather to multiple sets of variables

This function extends gather to work on multiple sets of columns.

gather_multi(data, key = "key", values = "values", varlist = list(), ...,
  key_func = NULL, na.rm = FALSE, convert = FALSE, factor_key = FALSE)

Arguments

data
A data frame.

key
See gather.

values
Multiple values as a character vector, or as unquoted using vars(). See gather.

varlist
A vector with elements that are created with vars. See details.

...
See gather.

key_func
A function to apply to the key variable. See details.

na.rm
In this setting it typically makes little sense to set this to TRUE. A warning will be provided. See gather.

convert
See gather.

factor_key
See gather.

Value

A data frame in so-called 'long' format with columns values.

Details

This function is an attempt to extend the tidyr gather function to deal with more than one set of inputs, which is very common in longitudinal design and survey data generally. It will return the same thing as gather but with extra columns for each of the values you're wishing to construct.

The values and varlist arguments must be of equal length. All of the following types of approaches will work:

values = vars(X, Y)

values = c('X', 'Y')

For those values, you could use the following:

varlist = vars(starts_with('X'), starts_with('Y'))

varlist = vars(c(X.1, X.2), c(Y.1, Y.2))

Technically, even this:

varlist = list(c('X.1', 'X.2'), c('Y.1', 'Y.2'))

But it's not recommended as you wouldn't have access to the tidyselect helper functions. It would even take integers but I'll not demonstrate poor programming practice.

However, the following would not work with the above values, because it has four elements instead of two:

varlist = vars(X.1, X.2, Y.1, Y.2)

Often the key is made up of variable names we may not want to use. The key_func argument can be used for this as a shortcut to a separate mutate step. It is up to you to create a function that does what you want.

Note

At present it will only keep the first 'key', as the rest are redundant. You can use the key_func argument to pretty it up.

Examples

library(tidyext); library(dplyr)

# example of longitudinal data with 4 waves
demo_data_wide = data.frame(id = 1:10,
                            X = matrix(rnorm(40), ncol = 4),
                            Y = matrix(sample(0:1, 40, replace = TRUE),
                            ncol = 4),
                            Z = matrix(rpois(40, 5), ncol = 4))

test <- gather_multi(demo_data_wide,
                     key = wave,
                     values = vars(X, Y, Z),
                     varlist = vars(starts_with('X'),
                                    starts_with('Y'),
                                    starts_with('Z')))
test

#>    id wave           X Y  Z
#> 1   1  X.1  1.07434588 0 10
#> 2   2  X.1 -0.66508825 1  1
#> 3   3  X.1  1.11395242 0  6
#> 4   4  X.1 -0.24589641 0  4
#> 5   5  X.1 -1.17756331 0  8
#> 6   6  X.1 -0.97585062 1  6
#> 7   7  X.1  1.06505732 0  6
#> 8   8  X.1  0.13167063 0  6
#> 9   9  X.1  0.48862881 0  5
#> 10 10  X.1 -1.69945057 1  3
#> 11  1  X.2 -1.47073631 0  5
#> 12  2  X.2  0.28415034 1  4
#> 13  3  X.2  1.33732041 0  6
#> 14  4  X.2  0.23669628 1  6
#> 15  5  X.2  1.31829338 0  4
#> 16  6  X.2  0.52390979 1  9
#> 17  7  X.2  0.60674805 1  4
#> 18  8  X.2 -0.10993567 1  2
#> 19  9  X.2  0.17218172 1  5
#> 20 10  X.2 -0.09032729 1  3
#> 21  1  X.3  1.92434334 0  3
#> 22  2  X.3  1.29839276 0  9
#> 23  3  X.3  0.74879127 0  7
#> 24  4  X.3  0.55622433 1  6
#> 25  5  X.3 -0.54825726 1  5
#> 26  6  X.3  1.11053489 1  5
#> 27  7  X.3 -2.61233433 0  2
#> 28  8  X.3 -0.15569378 1  4
#> 29  9  X.3  0.43388979 1  3
#> 30 10  X.3 -0.38195111 1  6
#> 31  1  X.4  0.42418757 1  6
#> 32  2  X.4  1.06310200 0  9
#> 33  3  X.4  1.04871262 1  3
#> 34  4  X.4 -0.03810289 0 10
#> 35  5  X.4  0.48614892 0  4
#> 36  6  X.4  1.67288261 1  6
#> 37  7  X.4 -0.35436116 1  7
#> 38  8  X.4  0.94634789 0  2
#> 39  9  X.4  1.31682636 1  5
#> 40 10  X.4 -0.29664002 1  7

test <- gather_multi(demo_data_wide,
                     key = wave,
                     values = c('X', 'Y', 'Z'),
                     varlist = vars(starts_with('X'),
                                    starts_with('Y'),
                                    starts_with('Z')),
                     key_func = function(x) substr(x, start=3, stop=3))
test

#>    id wave           X Y  Z
#> 1   1    1  1.07434588 0 10
#> 2   2    1 -0.66508825 1  1
#> 3   3    1  1.11395242 0  6
#> 4   4    1 -0.24589641 0  4
#> 5   5    1 -1.17756331 0  8
#> 6   6    1 -0.97585062 1  6
#> 7   7    1  1.06505732 0  6
#> 8   8    1  0.13167063 0  6
#> 9   9    1  0.48862881 0  5
#> 10 10    1 -1.69945057 1  3
#> 11  1    2 -1.47073631 0  5
#> 12  2    2  0.28415034 1  4
#> 13  3    2  1.33732041 0  6
#> 14  4    2  0.23669628 1  6
#> 15  5    2  1.31829338 0  4
#> 16  6    2  0.52390979 1  9
#> 17  7    2  0.60674805 1  4
#> 18  8    2 -0.10993567 1  2
#> 19  9    2  0.17218172 1  5
#> 20 10    2 -0.09032729 1  3
#> 21  1    3  1.92434334 0  3
#> 22  2    3  1.29839276 0  9
#> 23  3    3  0.74879127 0  7
#> 24  4    3  0.55622433 1  6
#> 25  5    3 -0.54825726 1  5
#> 26  6    3  1.11053489 1  5
#> 27  7    3 -2.61233433 0  2
#> 28  8    3 -0.15569378 1  4
#> 29  9    3  0.43388979 1  3
#> 30 10    3 -0.38195111 1  6
#> 31  1    4  0.42418757 1  6
#> 32  2    4  1.06310200 0  9
#> 33  3    4  1.04871262 1  3
#> 34  4    4 -0.03810289 0 10
#> 35  5    4  0.48614892 0  4
#> 36  6    4  1.67288261 1  6
#> 37  7    4 -0.35436116 1  7
#> 38  8    4  0.94634789 0  2
#> 39  9    4  1.31682636 1  5
#> 40 10    4 -0.29664002 1  7