Spread without frustration

Spread without frustration

spread2(data, key, value, compact = TRUE, fill = NA, convert = FALSE,
  drop = TRUE, sep = NULL)

Arguments

data
See spread.

key
See spread.

value
See spread.

compact
When you don't have a unique identifier there is more than one way the output could be produced. Actually two. One is rather messy but may in fact be what is desired. The other is more 'compact'. See the examples.

fill
See spread.

convert
See spread.

drop
See spread.

sep
See spread.

Value

A data frame with 'wide' format.

Details

Anyone that uses spread more than once has likely come across the duplicate identifiers problem. This function works just like select (it's mostly just a wrapper) except that it works when you don't have unique identifiers, which is actually quite often. The only additional functionality regards the compact argument. See the examples for what this means.

For more, see StackOverflow for example, or the numerous repeated issue brought up at the tidyr GitHub issues page. Search for 'spread duplicate' for all, not just open issues (direct link couldn't be handled by R documentation).

See also

Examples

library(tidyext); library(tidyr)

# initial example from spread
stocks_init <- data.frame(
time = as.Date('2009-01-01') + 0:9,
X = rnorm(10, 0, 1),
Y = rnorm(10, 0, 2),
Z = rnorm(10, 0, 4)
)

# a very common situation
stocks <- data.frame(
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4)
)

stocksm_init <- stocks_init %>% gather(stock, price, -time)
stocksm_init %>% spread(stock, price)    # works fine

#>          time            X           Y           Z
#> 1  2009-01-01 -0.098550690  3.68063348 -3.84628597
#> 2  2009-01-02  0.034766060 -1.40963933  4.06996421
#> 3  2009-01-03  0.386127022  0.01702062 -5.98421497
#> 4  2009-01-04  0.020831228  4.06837977 -4.73927492
#> 5  2009-01-05  0.007586777 -2.68337214  2.52093749
#> 6  2009-01-06  0.930844030  2.31795836  8.40501006
#> 7  2009-01-07 -0.684749941 -0.40641792 -2.45494724
#> 8  2009-01-08  0.337401513 -0.75605711 -6.53855309
#> 9  2009-01-09 -0.412137704  3.47222209 -0.04176447
#> 10 2009-01-10  0.934261130 -1.69049563 -2.62602456

stocksm <- stocks %>% gather(stock, price)
# stocksm %>% spread(stock, price)       # annoying
stocksm %>% spread2(stock, price)        # works fine

#>             X          Y          Z
#> 1  -0.6695334  1.4819334  2.2648065
#> 2  -0.4785890 -1.9948862  4.6088478
#> 3   1.3194563 -5.8779551 -3.0247895
#> 4   0.6365628  1.4380313 -1.9570333
#> 5   0.5143278 -1.3960101 -4.6642093
#> 6  -1.7513751 -3.7882517 -1.9186758
#> 7   0.8935975  0.1525985  0.4613929
#> 8   0.2230384  1.7506170 -7.0721936
#> 9   0.5808166  0.9076548 -5.6305557
#> 10 -0.1778214 -1.7014338  2.8367138

stocksm %>% spread2(stock, price, compact = FALSE)

#>             X          Y          Z
#> 1  -0.6695334         NA         NA
#> 2  -0.4785890         NA         NA
#> 3   1.3194563         NA         NA
#> 4   0.6365628         NA         NA
#> 5   0.5143278         NA         NA
#> 6  -1.7513751         NA         NA
#> 7   0.8935975         NA         NA
#> 8   0.2230384         NA         NA
#> 9   0.5808166         NA         NA
#> 10 -0.1778214         NA         NA
#> 11         NA  1.4819334         NA
#> 12         NA -1.9948862         NA
#> 13         NA -5.8779551         NA
#> 14         NA  1.4380313         NA
#> 15         NA -1.3960101         NA
#> 16         NA -3.7882517         NA
#> 17         NA  0.1525985         NA
#> 18         NA  1.7506170         NA
#> 19         NA  0.9076548         NA
#> 20         NA -1.7014338         NA
#> 21         NA         NA  2.2648065
#> 22         NA         NA  4.6088478
#> 23         NA         NA -3.0247895
#> 24         NA         NA -1.9570333
#> 25         NA         NA -4.6642093
#> 26         NA         NA -1.9186758
#> 27         NA         NA  0.4613929
#> 28         NA         NA -7.0721936
#> 29         NA         NA -5.6305557
#> 30         NA         NA  2.8367138

# works with NA
stocksm$price[sample(1:nrow(stocksm), 5)] = NA
stocksm %>% spread2(stock, price)

#>             X          Y         Z
#> 1  -0.6695334  1.4819334  2.264806
#> 2  -0.4785890 -1.9948862  4.608848
#> 3   1.3194563 -5.8779551 -3.024790
#> 4   0.6365628  1.4380313        NA
#> 5   0.5143278 -1.3960101 -4.664209
#> 6  -1.7513751         NA -1.918676
#> 7   0.8935975         NA        NA
#> 8   0.2230384         NA -7.072194
#> 9   0.5808166  0.9076548 -5.630556
#> 10 -0.1778214 -1.7014338  2.836714

stocksm %>% spread2(stock, price, compact = FALSE)

#>             X          Y         Z
#> 1  -0.6695334         NA        NA
#> 2  -0.4785890         NA        NA
#> 3   1.3194563         NA        NA
#> 4   0.6365628         NA        NA
#> 5   0.5143278         NA        NA
#> 6  -1.7513751         NA        NA
#> 7   0.8935975         NA        NA
#> 8   0.2230384         NA        NA
#> 9   0.5808166         NA        NA
#> 10 -0.1778214         NA        NA
#> 11         NA  1.4819334        NA
#> 12         NA -1.9948862        NA
#> 13         NA -5.8779551        NA
#> 14         NA  1.4380313        NA
#> 15         NA -1.3960101        NA
#> 16         NA         NA        NA
#> 17         NA         NA        NA
#> 18         NA         NA        NA
#> 19         NA  0.9076548        NA
#> 20         NA -1.7014338        NA
#> 21         NA         NA  2.264806
#> 22         NA         NA  4.608848
#> 23         NA         NA -3.024790
#> 24         NA         NA        NA
#> 25         NA         NA -4.664209
#> 26         NA         NA -1.918676
#> 27         NA         NA        NA
#> 28         NA         NA -7.072194
#> 29         NA         NA -5.630556
#> 30         NA         NA  2.836714