Spread without frustration

Spread without frustration

spread2(data, key, value, compact = TRUE, fill = NA, convert = FALSE,
  drop = TRUE, sep = NULL)

Arguments

data
See spread.

key
See spread.

value
See spread.

compact
When you don't have a unique identifier there is more than one way the output could be produced. Actually two. One is rather messy but may in fact be what is desired. The other is more 'compact'. See the examples.

fill
See spread.

convert
See spread.

drop
See spread.

sep
See spread.

Value

A data frame with 'wide' format.

Details

Anyone that uses spread more than once has likely come across the duplicate identifiers problem. This function works just like select (it's mostly just a wrapper) except that it works when you don't have unique identifiers, which is actually quite often. The only additional functionality regards the compact argument. See the examples for what this means.

For more, see StackOverflow for example, or the numerous repeated issue brought up at the tidyr GitHub issues page. Search for 'spread duplicate' for all, not just open issues (direct link couldn't be handled by R documentation).

See also

Examples

library(tidyext); library(tidyr)

# initial example from spread
stocks_init <- data.frame(
time = as.Date('2009-01-01') + 0:9,
X = rnorm(10, 0, 1),
Y = rnorm(10, 0, 2),
Z = rnorm(10, 0, 4)
)

# a very common situation
stocks <- data.frame(
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4)
)

stocksm_init <- stocks_init %>% gather(stock, price, -time)
stocksm_init %>% spread(stock, price)    # works fine

#>          time          X          Y          Z
#> 1  2009-01-01  0.3224827  0.9614505  4.7263023
#> 2  2009-01-02  0.3666744  3.2088147  0.7935684
#> 3  2009-01-03  1.1298352 -3.0300491 -1.6016210
#> 4  2009-01-04 -0.9414981 -2.8320478  2.4646171
#> 5  2009-01-05  0.2178376  1.7535547  7.8966270
#> 6  2009-01-06  1.4154123  1.2482648  7.5386493
#> 7  2009-01-07 -0.3837330  4.2245546 -6.3544822
#> 8  2009-01-08 -0.1740864 -0.7122488 -2.1596927
#> 9  2009-01-09 -0.2217445 -2.1289284 -4.6778459
#> 10 2009-01-10 -1.0095287  2.1542331  2.2364240

stocksm <- stocks %>% gather(stock, price)
# stocksm %>% spread(stock, price)       # annoying
stocksm %>% spread2(stock, price)        # works fine

#>              X          Y          Z
#> 1  -1.81934725  0.1440514 -0.6771203
#> 2   0.39334397  4.2528891  1.1801190
#> 3   0.04213411 -2.9523938  5.0653623
#> 4   1.17966418  0.8157770 -4.5413730
#> 5  -0.25692118  2.7879556 -4.5242152
#> 6  -1.05633610  0.7205566  0.4399735
#> 7   0.19877721  1.3091005  3.4116216
#> 8   0.65053355  2.1043108 -0.9373514
#> 9   0.34391334 -3.9591103  8.3467543
#> 10  1.47753231  2.4167712 -0.4436775

stocksm %>% spread2(stock, price, compact = FALSE)

#>              X          Y          Z
#> 1  -1.81934725         NA         NA
#> 2   0.39334397         NA         NA
#> 3   0.04213411         NA         NA
#> 4   1.17966418         NA         NA
#> 5  -0.25692118         NA         NA
#> 6  -1.05633610         NA         NA
#> 7   0.19877721         NA         NA
#> 8   0.65053355         NA         NA
#> 9   0.34391334         NA         NA
#> 10  1.47753231         NA         NA
#> 11          NA  0.1440514         NA
#> 12          NA  4.2528891         NA
#> 13          NA -2.9523938         NA
#> 14          NA  0.8157770         NA
#> 15          NA  2.7879556         NA
#> 16          NA  0.7205566         NA
#> 17          NA  1.3091005         NA
#> 18          NA  2.1043108         NA
#> 19          NA -3.9591103         NA
#> 20          NA  2.4167712         NA
#> 21          NA         NA -0.6771203
#> 22          NA         NA  1.1801190
#> 23          NA         NA  5.0653623
#> 24          NA         NA -4.5413730
#> 25          NA         NA -4.5242152
#> 26          NA         NA  0.4399735
#> 27          NA         NA  3.4116216
#> 28          NA         NA -0.9373514
#> 29          NA         NA  8.3467543
#> 30          NA         NA -0.4436775

# works with NA
stocksm$price[sample(1:nrow(stocksm), 5)] = NA
stocksm %>% spread2(stock, price)

#>             X          Y          Z
#> 1  -1.8193472         NA -0.6771203
#> 2   0.3933440  4.2528891  1.1801190
#> 3          NA -2.9523938  5.0653623
#> 4          NA  0.8157770 -4.5413730
#> 5  -0.2569212  2.7879556         NA
#> 6  -1.0563361  0.7205566  0.4399735
#> 7   0.1987772  1.3091005  3.4116216
#> 8   0.6505336  2.1043108 -0.9373514
#> 9   0.3439133 -3.9591103  8.3467543
#> 10  1.4775323         NA -0.4436775

stocksm %>% spread2(stock, price, compact = FALSE)

#>             X          Y          Z
#> 1  -1.8193472         NA         NA
#> 2   0.3933440         NA         NA
#> 3          NA         NA         NA
#> 4          NA         NA         NA
#> 5  -0.2569212         NA         NA
#> 6  -1.0563361         NA         NA
#> 7   0.1987772         NA         NA
#> 8   0.6505336         NA         NA
#> 9   0.3439133         NA         NA
#> 10  1.4775323         NA         NA
#> 11         NA         NA         NA
#> 12         NA  4.2528891         NA
#> 13         NA -2.9523938         NA
#> 14         NA  0.8157770         NA
#> 15         NA  2.7879556         NA
#> 16         NA  0.7205566         NA
#> 17         NA  1.3091005         NA
#> 18         NA  2.1043108         NA
#> 19         NA -3.9591103         NA
#> 20         NA         NA         NA
#> 21         NA         NA -0.6771203
#> 22         NA         NA  1.1801190
#> 23         NA         NA  5.0653623
#> 24         NA         NA -4.5413730
#> 25         NA         NA         NA
#> 26         NA         NA  0.4399735
#> 27         NA         NA  3.4116216
#> 28         NA         NA -0.9373514
#> 29         NA         NA  8.3467543
#> 30         NA         NA -0.4436775