Spread without frustration

Spread without frustration

spread2(data, key, value, compact = TRUE, fill = NA, convert = FALSE,
  drop = TRUE, sep = NULL)

Arguments

data
See spread.

key
See spread.

value
See spread.

compact
When you don't have a unique identifier there is more than one way the output could be produced. Actually two. One is rather messy but may in fact be what is desired. The other is more 'compact'. See the examples.

fill
See spread.

convert
See spread.

drop
See spread.

sep
See spread.

Value

A data frame with 'wide' format.

Details

Anyone that uses spread more than once has likely come across the duplicate identifiers problem. This function works just like select (it's mostly just a wrapper) except that it works when you don't have unique identifiers, which is actually quite often. The only additional functionality regards the compact argument. See the examples for what this means.

For more, see StackOverflow for example, or the numerous repeated issue brought up at the tidyr GitHub issues page. Search for 'spread duplicate' for all, not just open issues (direct link couldn't be handled by R documentation).

See also

Examples

library(tidyext); library(tidyr)

# initial example from spread
stocks_init <- data.frame(
time = as.Date('2009-01-01') + 0:9,
X = rnorm(10, 0, 1),
Y = rnorm(10, 0, 2),
Z = rnorm(10, 0, 4)
)

# a very common situation
stocks <- data.frame(
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4)
)

stocksm_init <- stocks_init %>% gather(stock, price, -time)
stocksm_init %>% spread(stock, price)    # works fine

#>          time          X          Y           Z
#> 1  2009-01-01  0.6884226  2.0017709  1.29954048
#> 2  2009-01-02  0.3243352  0.6263323 -4.23046207
#> 3  2009-01-03 -0.1388366 -1.9610601 -5.46222900
#> 4  2009-01-04 -1.1836451 -2.1315168  1.42190664
#> 5  2009-01-05  1.6123666 -2.0279735 -2.69137612
#> 6  2009-01-06 -0.6431638 -2.7332627  1.61258387
#> 7  2009-01-07 -1.1657161 -3.5350091  1.33316664
#> 8  2009-01-08  0.3135301 -1.1303017  5.91156447
#> 9  2009-01-09  0.5484061 -0.5170878 -0.04451998
#> 10 2009-01-10 -1.2573885  3.2624551 -2.71390525

stocksm <- stocks %>% gather(stock, price)
# stocksm %>% spread(stock, price)       # annoying
stocksm %>% spread2(stock, price)        # works fine

#>              X          Y          Z
#> 1  -0.35856540 -1.3883178  6.1018751
#> 2   0.60161662  0.2630438 -2.2582307
#> 3  -0.75048498  0.6555027  1.1851909
#> 4  -0.39653934  2.8723032 -0.9968521
#> 5  -1.92652463 -2.6574446 -0.8283984
#> 6  -2.20590368  0.3514273  3.2882265
#> 7   1.43222841  1.0532184  2.8145529
#> 8  -0.23972892 -0.3449991 -0.6308590
#> 9  -0.02936022  0.3353793 -0.7072142
#> 10 -2.81003183  2.3530661  5.6901434

stocksm %>% spread2(stock, price, compact = FALSE)

#>              X          Y          Z
#> 1  -0.35856540         NA         NA
#> 2   0.60161662         NA         NA
#> 3  -0.75048498         NA         NA
#> 4  -0.39653934         NA         NA
#> 5  -1.92652463         NA         NA
#> 6  -2.20590368         NA         NA
#> 7   1.43222841         NA         NA
#> 8  -0.23972892         NA         NA
#> 9  -0.02936022         NA         NA
#> 10 -2.81003183         NA         NA
#> 11          NA -1.3883178         NA
#> 12          NA  0.2630438         NA
#> 13          NA  0.6555027         NA
#> 14          NA  2.8723032         NA
#> 15          NA -2.6574446         NA
#> 16          NA  0.3514273         NA
#> 17          NA  1.0532184         NA
#> 18          NA -0.3449991         NA
#> 19          NA  0.3353793         NA
#> 20          NA  2.3530661         NA
#> 21          NA         NA  6.1018751
#> 22          NA         NA -2.2582307
#> 23          NA         NA  1.1851909
#> 24          NA         NA -0.9968521
#> 25          NA         NA -0.8283984
#> 26          NA         NA  3.2882265
#> 27          NA         NA  2.8145529
#> 28          NA         NA -0.6308590
#> 29          NA         NA -0.7072142
#> 30          NA         NA  5.6901434

# works with NA
stocksm$price[sample(1:nrow(stocksm), 5)] = NA
stocksm %>% spread2(stock, price)

#>              X          Y          Z
#> 1  -0.35856540 -1.3883178  6.1018751
#> 2   0.60161662  0.2630438 -2.2582307
#> 3           NA  0.6555027  1.1851909
#> 4  -0.39653934         NA -0.9968521
#> 5  -1.92652463 -2.6574446 -0.8283984
#> 6           NA  0.3514273  3.2882265
#> 7   1.43222841  1.0532184  2.8145529
#> 8  -0.23972892 -0.3449991         NA
#> 9  -0.02936022  0.3353793 -0.7072142
#> 10 -2.81003183         NA  5.6901434

stocksm %>% spread2(stock, price, compact = FALSE)

#>              X          Y          Z
#> 1  -0.35856540         NA         NA
#> 2   0.60161662         NA         NA
#> 3           NA         NA         NA
#> 4  -0.39653934         NA         NA
#> 5  -1.92652463         NA         NA
#> 6           NA         NA         NA
#> 7   1.43222841         NA         NA
#> 8  -0.23972892         NA         NA
#> 9  -0.02936022         NA         NA
#> 10 -2.81003183         NA         NA
#> 11          NA -1.3883178         NA
#> 12          NA  0.2630438         NA
#> 13          NA  0.6555027         NA
#> 14          NA         NA         NA
#> 15          NA -2.6574446         NA
#> 16          NA  0.3514273         NA
#> 17          NA  1.0532184         NA
#> 18          NA -0.3449991         NA
#> 19          NA  0.3353793         NA
#> 20          NA         NA         NA
#> 21          NA         NA  6.1018751
#> 22          NA         NA -2.2582307
#> 23          NA         NA  1.1851909
#> 24          NA         NA -0.9968521
#> 25          NA         NA -0.8283984
#> 26          NA         NA  3.2882265
#> 27          NA         NA  2.8145529
#> 28          NA         NA         NA
#> 29          NA         NA -0.7072142
#> 30          NA         NA  5.6901434